+++ hash.h
@@ -38,7 +38,7 @@ extern uint_fast64_t hash_material_key (const board_t * board);
 extern uint_fast64_t hash_piece_key    (int_fast32_t piece, int_fast32_t square);
 extern uint_fast64_t hash_castle_key   (int_fast32_t flags);
 extern uint_fast64_t hash_ep_key       (int_fast32_t square);
-extern uint_fast64_t hash_turn_key     (int_fast32_t colour);
+extern uint_fast64_t hash_turn_key     (int_fast8_t colour);
 
 #endif // !defined HASH_H
